% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{AddTrailingSlash}
\alias{AddTrailingSlash}
\title{AddTrailingSlash}
\usage{
AddTrailingSlash(string = InputD)
}
\arguments{
\item{string}{The file path potentially missing the trailing slash}
}
\description{
Add final '/' if missing from a file path.
}
\examples{
AddTrailingSlash (string = "stairway/to/heaven")
}
