% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{as.logical.wNames}
\alias{as.logical.wNames}
\title{as.logical.wNames}
\usage{
\method{as.logical}{wNames}(x, ...)
}
\arguments{
\item{x}{A vector with names that will be converted to a logical vector}

\item{...}{Pass any other argument.}
}
\description{
Converts your input vector into a logical vector, and puts the original character values
into the names of the new vector, unless it already has names.
}
\examples{
# x= -1:2; names(x)=LETTERS[1:4]; as.logical.wNames(x)
}
