% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{Color_Check}
\alias{Color_Check}
\title{Color_Check}
\usage{
Color_Check(..., incrBottMarginBy = 0, savefile = FALSE)
}
\arguments{
\item{...}{Additional parameters.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}
}
\description{
Display the colors encoded by the numbers / color-ID-s you pass on to this function
}
\examples{
Color_Check(1:3)
}
