% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{ParseFilePath}
\alias{ParseFilePath}
\title{ParseFilePath}
\usage{
ParseFilePath(...)
}
\arguments{
\item{string}{The file path potentially having Double Slash}
}
\description{
ParseFilePath pastes elements by slash, then removes Double Slashes '//' from a string and adds a final '/' if missing from a file path.
}
\examples{
ParseFilePath(string = "stairway///to/heaven")
}
