% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{RemoveDoubleSlash}
\alias{RemoveDoubleSlash}
\title{RemoveDoubleSlash}
\usage{
RemoveDoubleSlash(string = "stairway//to/heaven")
}
\arguments{
\item{string}{The file path potentially having Double Slash}
}
\description{
Remove Double Slash '/'  from string.
}
\examples{
RemoveDoubleSlash (string = "stairway//to/heaven")
}
