% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReportsDev.R
\name{unique.wNames}
\alias{unique.wNames}
\title{unique.wNames}
\usage{
\method{unique}{wNames}(x, ...)
}
\arguments{
\item{x}{A vector with values and names}

\item{...}{Pass any other argument to}
}
\description{
Get the unique elements from a vector keeping names (of the first elements of each category).
}
\examples{
x <- c(1:4, 1:2);  names(x) = 6:1; unique.wNames (x)
}
